requirements.txt
# Core LLM & AI
google-genai

# Document Parsing
docling>=2.0.0

# Vector Store
chromadb>=0.5.0

# Evaluation
ragas>=0.1.0
datasets>=2.0.0

# Prompt templates
pyyaml>=6.0

# Environment management
python-dotenv>=1.0.0

# CLI & Display
typer>=0.12.0
rich>=13.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0
